% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bMCW_approximated_test.R
\name{bMCW_approximated_test}
\alias{bMCW_approximated_test}
\title{Biased-Measure Monte Carlo-Wilcoxon (bMCW) Approximated Test}
\usage{
bMCW_approximated_test(x)
}
\arguments{
\item{x}{Data for bMCW approximated tests provided by function \emph{bMCWtest}.}
}
\description{
Biased-Measure Monte Carlo-Wilcoxon (bMCW) Approximated Test
}
\keyword{internal}
