% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bMCW_approximated_test_simulated_BIs.R
\name{bMCW_approximated_test_simulated_BIs}
\alias{bMCW_approximated_test_simulated_BIs}
\title{Biased-Measure Monte Carlo-Wilcoxon (bMCW) Approximated Test: Simulated Bias Indexes (BIs)}
\usage{
bMCW_approximated_test_simulated_BIs(x)
}
\arguments{
\item{x}{Data for bMCW approximated tests provided by function \emph{bMCW_approximated_test}.}
}
\description{
Biased-Measure Monte Carlo-Wilcoxon (bMCW) Approximated Test: Simulated Bias Indexes (BIs)
}
\keyword{internal}
