% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bMCW_exact_test.R
\name{bMCW_exact_test}
\alias{bMCW_exact_test}
\title{Biased-Measure Monte Carlo-Wilcoxon (bMCW) Exact Test}
\usage{
bMCW_exact_test(x)
}
\arguments{
\item{x}{Data for bMCW exact tests provided by function \emph{bMCWtest}.}
}
\description{
Biased-Measure Monte Carlo-Wilcoxon (bMCW) Exact Test
}
\keyword{internal}
