% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bMCW_exact_test_subset_combinations.R
\name{bMCW_exact_test_subset_combinations}
\alias{bMCW_exact_test_subset_combinations}
\title{Biased-Measure Monte Carlo-Wilcoxon (bMCW) Exact Test: All Subset Combinations}
\usage{
bMCW_exact_test_subset_combinations(x)
}
\arguments{
\item{x}{Data for bMCW exact tests provided by function \emph{bMCW_exact_test}.}
}
\description{
Biased-Measure Monte Carlo-Wilcoxon (bMCW) Exact Test: All Subset Combinations
}
\keyword{internal}
