% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bMCW_exact_test_wholeset_combinations.R
\name{bMCW_exact_test_wholeset_combinations}
\alias{bMCW_exact_test_wholeset_combinations}
\title{Biased-Measure Monte Carlo-Wilcoxon (bMCW) Exact Test: All Whole-set Combinations}
\usage{
bMCW_exact_test_wholeset_combinations(x)
}
\arguments{
\item{x}{Data for bMCW exact tests provided by function \emph{bMCW_exact_test}.}
}
\description{
Biased-Measure Monte Carlo-Wilcoxon (bMCW) Exact Test: All Whole-set Combinations
}
\keyword{internal}
