% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bMCWtest.R
\name{bMCWtest}
\alias{bMCWtest}
\title{Bias-measures Monte Carlo-Wilcoxon (bMCW) Test}
\format{
When executing the \emph{bMCWtest} function, users must provide the path to a local CSV file named \emph{X_bMCWtest_data.csv}, where \emph{X} serves as a user-defined identifier. \emph{X_bMCWtest_data.csv} should include the following columns:
\itemize{
\item Column \emph{bias_value} contains the value of the bias measure under analysis.
\item Columns \emph{subset_x}, where \emph{x} represents the specific type of subset for each column, such as "chr" for chromosomes or "GO" for Gene Ontology. These columns are required if users intend to assess whether bias measures for certain subsets of elements in the dataset are significantly biased in the same direction. Columns \emph{subset_x} can indicate whether an element belongs to a subset using either "YES" and "NO", or specific subset names like "chr1" or "chrX", or a combination of both, such as "chr1", "chrX" and "NO". The function \emph{bMCW test} will transform the dataset to conduct independent analysis of each subset of elements marked as "YES" or with a specific subset name in each \emph{subset_x} column.
\item As many informative columns as needed by users to contextualize the results of each test. The names of these columns should not contain the terms \emph{bias_value} or \emph{subset}. While these columns are optional when running a single test, at least one column is required when running multiple tests simultaneously. All rows for each individual test must contain the same information in these columns.
\item Users can specify columns with information relevant about each element or row using the column name structure \emph{element_x}, where \emph{x} indicates the specific information in each column (see example). However, \emph{element_x} columns are not essential for bMCW testing and will not be included in the results file.
}
}
\usage{
bMCWtest(path, max_rearrangements)
}
\arguments{
\item{path}{Path for the local CSV file containing the entry dataset formatted for bMCW tests.}

\item{max_rearrangements}{User-defined maximum number of rearrangements of the dataset used by the function \emph{bMCWtest} to generate a collection of expected-by-chance bMCW_wBIs and bMCW_sBIs and estimate the statistical significance of observed bMCW_BIs. If the number of distinct dataset rearrangements is less than \emph{max_rearrangements}, \emph{bMCWtest} calculates bMCW_wBIs and bMCW_sBIs for all possible data rearrangements. If the number of distinct dataset rearrangements is greater than \emph{max_rearrangements}, \emph{bMCWtest} will perform N = \emph{max_rearrangements} random measure rearrangements to calculate the collection of expected-by-chance bMCW_wBIs and bMCW_sBIs.}
}
\value{
The \emph{bMCWtest} function reports to the console the total number of tests it will execute, and their exact and approximated counts. It also creates a CSV file named \emph{X_bMCWtest_results.csv}, where \emph{X} is a user-defined identifier for the entry dataset CSV file. The \emph{X_bMCWtest_results.csv} file contains one row for each bMCWtest to indicate the results of whole-set bMCW testing, and as many rows as necessary to indicate the results of subset bMCW testing. Rows for whole-set analyses will be at the top of \emph{X_bMCWtest_results.csv} file. The \emph{X_bMCWtest_results.csv} file includes the following columns:
\itemize{
\item User-provided informative columns to contextualize the results of each test.
\item Column \emph{subset_type} indicates whether the results in each row corresponds to \emph{whole-set} tests or specific \emph{subset} tests, such as "chr" for chromosomes or "GO" for Gene Ontology terms.
\item Column \emph{tested_subset} indicates the name of the subset under analysis. For \emph{whole-set} tests, the \emph{tested_subset} column indicates "none". For subset tests, the \emph{tested_subset} column indicates "YES" or the specific name of the subset under analysis, such as "chr1" or "chrX".
\item Columns \emph{N} and \emph{n} indicate the total number of elements in the whole set and those associated with the subset under analysis, respectively, after removing missing values (NAs). For \emph{whole-set} tests, columns \emph{N} and \emph{n} have the same value.
\item Column \emph{test_type} distinguishes between exact and approximated tests.
\item Column \emph{BI_type} indicates whether results correspond to \emph{whole-set} tests (bMCW_wBI) or to \emph{subset} tests (bMCW_sBIs).
\item Column \emph{observed_BI} contains the value of bMCW_BIs obtained from analyzing the user-provided dataset.
\item Column \emph{expected_by_chance_BI_N} indicates the number of data rearrangements used to calculate the expected-by-chance bMCW_wBIs and bMCW_sBIs. This value corresponds to the lowest number between all possible measure rearrangements and the parameter \emph{max_rearrangements}.
\item Columns \emph{pupper} and \emph{plower} represent the \emph{P~upper~} and \emph{P~lower~} values, respectively. They denote the fraction of expected-by-chance bMCW_wBIs or bMCW_sBIs with values higher or equal to and lower or equal to the observed bMCW_wBIs or bMCW_sBIs, respectively.
}
}
\description{
The bMCW test is a statistical tool to assess whether a set of measures of bias for a quantitative trait between two conditions or a subset of these bias measures are themselves significantly biased in the same direction. For instance, bMCW tests can be used to analyze bias indexes obtained using other MCW tests or fold change for transcript abundances spanning the entire transcriptome or only for genes located in specific genomic regions from two sets of mice exposed to different conditions.
}
\details{
The function \emph{bMCWtest} eliminates missing values (NAs) from the dataset before proceeding with the following steps.
\itemize{
\item To estimate the bias for all bias measures in the entire dataset or a subset of them, the function \emph{bMCWtest} performs the following tasks:
\itemize{
\item It ranks all bias measures with non-zero values from lowest to highest. Bias measures with a value of 0 are assigned a 0 rank. If multiple bias measures have the same absolute value, all tied bias measures are asssigned the lowest rank possible.
\item It assigns each rank a sign based on the sign of its corresponding bias measure.
\item It calculates a whole-set bias index (bMCW_wBI) by summing the signed ranks for all elements in the dataset and dividing it by the maximum number that sum could have if all bias measures were positive. Consequently, bMCW_wBI ranges between 1 when all bias measures are positive, and -1 when all bias measures are negative.
\item It calculates a subset bias index (bMCW_sBI) for each subset of elements under analysis by summing the signed ranks for the elements in the subset and dividing it by the maximum number that sum could have if the elements in the subset had the highest possible positive bias measures. Consequently, bMCW_sBI ranges between 1 when the bias measures for the subset in question have the highest positive bias measures in the entire dataset, and -1 when the bias measures for the subset in question have the lowest negative bias measures in the entire dataset.
}
\item To assess the significance of the bMCW-wBIs and bMCW-sBIs obtained from the user-provided dataset (observed bMCW-wBIs and bMCW-sBIs), the function \emph{bMCWtest} performs the following tasks,
\itemize{
\item It generates a collection of expected-by-chance bMCW_wBIs by rearranging the signs of all signed ranks multiple times. The function \emph{bMCWtest} also generates a collection of expected-by-chance bMCW_sBIs by rearranging the subset of elements multiple times. The user-provided parameter \emph{max_rearrangements} determines the two paths that the function \emph{bMCWtest} can follow to generate the collection of expected-by-chance bMCW_wBIs and bMCW_sBIs:
\itemize{
\item \emph{bMCW exact testing}: If the number of distinct bias measure rearrangements that can alter their initial sign distribution or subset distribution is less than \emph{max_rearrangements}, the function \emph{bMCWtest} calculates bMCW_wBIs or bMCW_sBIs for all possible data rearrangements.
\item \emph{bMCW approximated testing}: If the number of distinct bias measure rearrangements that can alter their initial sign distribution or subset distribution is greater than \emph{max_rearrangements}, the function \emph{bMCWtest} performs N = \emph{max_rearrangements} random measure rearrangements to calculate the collection of expected-by-chance bMCW_wBIs or bMCW_sBIs.
}
\item It calculates \emph{P~upper~} and \emph{P~lower~} values, as the fraction of expected-by-chance bMCW-wBIs and bMCW-sBIs that are higher or equal to and lower or equal to the observed bMCW-wBIs and bMCW-sBIs, respectively.
}
}
}
\examples{
# running bMCWtest with a vertical entry dataset
path <- system.file("extdata", "bMCWtest_example_data.csv", package = "MCWtests")
bMCWtest_results <- bMCWtest(path, 200)

}
