% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbMCW_approximated_test.R
\name{mbMCW_approximated_test}
\alias{mbMCW_approximated_test}
\title{Matched-Pairs Bivariate Monte Carlo-Wilcoxon (mbMCW) Approximated Test}
\usage{
mbMCW_approximated_test(x)
}
\arguments{
\item{x}{Data for mbMCW approximated tests provided by function \emph{mbMCWtest}.}
}
\description{
Matched-Pairs Bivariate Monte Carlo-Wilcoxon (mbMCW) Approximated Test
}
\keyword{internal}
