% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbMCW_approximated_test_simulated_BIs.R
\name{mbMCW_approximated_test_simulated_BIs}
\alias{mbMCW_approximated_test_simulated_BIs}
\title{Matched-Pairs Bivariate Monte Carlo-Wilcoxon (mbMCW) Approximated Test: Simulated Bias Indexes (BIs)}
\usage{
mbMCW_approximated_test_simulated_BIs(x)
}
\arguments{
\item{x}{Data for mbMCW approximated tests provided by function \emph{mbMCW_approximated_test}.}
}
\description{
Matched-Pairs Bivariate Monte Carlo-Wilcoxon (mbMCW) Approximated Test: Simulated Bias Indexes (BIs)
}
\keyword{internal}
