% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbMCW_exact_test.R
\name{mbMCW_exact_test}
\alias{mbMCW_exact_test}
\title{Matched-Pairs Bivariate Monte Carlo-Wilcoxon (mbMCW) Exact Test}
\usage{
mbMCW_exact_test(x)
}
\arguments{
\item{x}{Data for mbMCW exact tests provided by function \emph{mbMCWtest}.}
}
\description{
Matched-Pairs Bivariate Monte Carlo-Wilcoxon (mbMCW) Exact Test
}
\keyword{internal}
