% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muMCW_approximated_test.R
\name{muMCW_approximated_test}
\alias{muMCW_approximated_test}
\title{Matched-Pairs Univariate Monte Carlo-Wilcoxon (muMCW) Approximated Test}
\usage{
muMCW_approximated_test(x)
}
\arguments{
\item{x}{Data for muMCW approximated tests provided by function \emph{muMCWtest}.}
}
\description{
Matched-Pairs Univariate Monte Carlo-Wilcoxon (muMCW) Approximated Test
}
\keyword{internal}
