% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muMCW_approximated_test_observed_BIs.R
\name{muMCW_approximated_test_observed_BIs}
\alias{muMCW_approximated_test_observed_BIs}
\title{Matched-Pairs Univariate Monte Carlo-Wilcoxon (muMCW) Approximated Test: Observed Bias Indexes (BIs)}
\usage{
muMCW_approximated_test_observed_BIs(x)
}
\arguments{
\item{x}{Data for muMCW approximated tests provided by function \emph{muMCW_approximated_test}.}
}
\description{
Matched-Pairs Univariate Monte Carlo-Wilcoxon (muMCW) Approximated Test: Observed Bias Indexes (BIs)
}
\keyword{internal}
