% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muMCW_exact_test.R
\name{muMCW_exact_test}
\alias{muMCW_exact_test}
\title{Matched-Pairs Univariate Monte Carlo-Wilcoxon (muMCW) Exact Test}
\usage{
muMCW_exact_test(x)
}
\arguments{
\item{x}{Data for muMCW exact tests provided by function \emph{muMCWtest}.}
}
\description{
Matched-Pairs Univariate Monte Carlo-Wilcoxon (muMCW) Exact Test
}
\keyword{internal}
