% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muMCW_exact_test_combinations.R
\name{muMCW_exact_test_combinations}
\alias{muMCW_exact_test_combinations}
\title{Matched-Pairs Univariate Monte Carlo-Wilcoxon (muMCW) Exact Test: Data Combinations}
\usage{
muMCW_exact_test_combinations(x)
}
\arguments{
\item{x}{Data for muMCW exact tests provided by function \emph{muMCW_exact_test}.}
}
\description{
Matched-Pairs Univariate Monte Carlo-Wilcoxon (muMCW) Exact Test: Data Combinations
}
\keyword{internal}
