% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muMCWtest.R
\name{muMCWtest}
\alias{muMCWtest}
\title{Matched-Measures Univariate Monte Carlo-Wilcoxon (muMCW) Test}
\format{
When executing the \emph{muMCWtest} function, users must provide the path to a local CSV file named \emph{X_muMCWtest_data.csv}, where \emph{X} serves as a user-defined identifier. \emph{X_muMCWtest_data.csv} can be structured in two distinct formats:
\itemize{
\item \strong{Vertical layout}: This format allows appending datasets with varying structures, such as different numbers of measure matched-pairs for each appended test. Vertical entry datasets should include the following columns:
\itemize{
\item Columns \emph{condition_a} and \emph{condition_b} uniquely identify the two conditions under which matched-paired measures were collected.
\item Columns \emph{value_a} and \emph{value_b} contain the actual measures under analysis.
\item As many informative columns as needed by users to contextualize the results of each test. The names of the these columns should not contain the terms \emph{condition} or \emph{value}. While these columns are optional when running a single test, at least one column is required when running multiple tests simultaneously. All rows for each individual test must contain the same information in these columns.
}
\item \strong{Horizontal layout}: This format allows appending datasets with similar structures, such as the same number of matched-paired measures for each appended test. Horizontal entry datasets should include the following columns:
\itemize{
\item Columns \emph{condition_a} and \emph{condition_b} uniquely identify the two conditions under which matched-paired measures were collected.
\item Columns \emph{a.i} and \emph{b.i}, where \emph{i} represents integers to differentiate each specific matched-pairs of measures, contain the actual measures under analysis.
\item As many informative columns as needed by users to contextualize the results of each test. The names of these columns should not contain the term \emph{condition} or have the same structure as the \emph{a.i} and \emph{b.i} columns. While these columns are optional when running a single test, at least one column is required when running multiple tests simultaneously.
}
}
}
\usage{
muMCWtest(path, max_rearrangements)
}
\arguments{
\item{path}{Path for the local CSV file containing the entry dataset formatted for muMCW tests.}

\item{max_rearrangements}{User-defined maximum number of rearrangements of the dataset used by the function \emph{muMCWtest} to generate a collection of expected-by-chance muMCW_BIs and estimate the statistical significance of observed muMCW_BIs. If the number of distinct dataset rearrangements is less than \emph{max_rearrangements}, \emph{muMCWtest} calculates muMCW_BIs for all possible data rearrangements. If the number of distinct dataset rearrangements is greater than \emph{max_rearrangements}, \emph{muMCWtest} will perform N = \emph{max_rearrangements} random measure rearrangements to calculate the collection of expected-by-chance muMCW_BIs.}
}
\value{
The \emph{muMCWtest} function reports to the console the total number of tests it will execute, and their exact and approximated counts. It also creates a CSV file named \emph{X_muMCWtest_results.csv} where \emph{X} is a user-defined identifier for the entry dataset CSV file. The \emph{X_muMCWtest_results.csv} file contains two rows for each muMCW test, with muMCW_BIs calculated for each possible condition contrast (\emph{e.g.}, \emph{a-b} and \emph{b-a}). The \emph{X_muMCWtest_results.csv} file includes the following columns:
\itemize{
\item User-provided informative columns to contextualize the results of each test.
\item Columns \emph{condition_a} and \emph{condition_b} indicate the two conditions for which matched-paired measures were provided.
\item Column \emph{N} indicates the total number of measure matched-pairs after removing matched-pairs with missing values (NAs).
\item Column \emph{test_type} distinguishes between exact and approximated tests.
\item Column \emph{BI_type} indicates muMCW_BI.
\item Column \emph{condition_contrast} indicates the condition contrast for each row of results.
\item Column \emph{observed_BI} contains the value of muMCW_BIs obtained from analyzing the user-provided dataset.
\item Column \emph{expected_by_chance_BI_N} indicates the number of data rearrangements used to calculate the expected-by-chance muMCW_BIs. This value corresponds to the lowest number between all possible measure rearrangements and the parameter \emph{max_rearrangements}.
\item Columns \emph{pupper} and \emph{plower} represent \emph{P~upper~} and \emph{P~lower~} values, respectively. They denote the fraction of expected-by-chance muMCW_BIs with values higher or equal to and lower or equal to the observed muMCW_BIs, respectively.
}
}
\description{
The muMCW test is a statistical tool to assess whether one set of inherently matched-paired measures is significantly biased in the same direction. For instance, muMCW tests can be used to analyze bodyweights or transcript abundances determined at two different timepoints for the same set of mice.
}
\details{
The function \emph{muMCWtest} eliminates any matched-paired measures with at least one missing value (NA) before proceeding with the following steps.
\itemize{
\item To estimate the bias for all matched-paired measures in the dataset, the function \emph{muMCWtest} performs the following tasks:
\itemize{
\item For each matched-pair of measures, it subtracts values for the two possible condition contrasts (\emph{e.g.}, \emph{a-b} and \emph{b-a}).
\item For each condition contrast, it ranks the absolute values of non-zero differences from lowest to highest. Measure pair differences with a value of 0 are assigned a 0 rank. If multiple measure pair differences have the same absolute value, all tied measure pair differences are assigned the lowest rank possible.
\item It assigns each measure pair rank a sign based on the sign of its corresponding measure pair difference.
\item It sums the signed ranks for each condition contrast.
\item It calculates muMCW_BI by dividing each sum of signed ranks by the maximum number that sum could have if the corresponding measure pairs had the highest possible positive ranks. Consequently, muMCW_BI ranges between 1 when all measures corresponding to the first condition are higher than all measures corresponding to the second condition, and -1 when all measures corresponding to the first condition are lower than all measures corresponding to the second condition.
}
\item To assess the significance of the muMCW_BIs obtained from the user-provided dataset (observed muMCW_BIs), the function \emph{muMCWtest} performs the following tasks:
\itemize{
\item It generates a collection of expected-by-chance muMCW_BIs. These expected values are obtained by rearranging the measures between and within the two conditions multiple times. The user-provided parameter \emph{max_rearrangements} determines the two paths that the function \emph{muMCWtest} can follow to generate the collection of expected-by-chance muMCW_BIs:
\itemize{
\item \emph{muMCW exact testing}: If the number of distinct measure rearrangements that can alter their initial pair and set distribution is less than \emph{max_rearrangements}, the function \emph{muMCWtest} calculates muMCW_BIs for all possible data rearrangements.
\item \emph{muMCW approximated testing}: If the number of distinct measure rearrangements that can alter their initial pair and set distribution is greater than \emph{max_rearrangements}, the function \emph{muMCWtest} performs N = \emph{max_rearrangements} random measure rearrangements to calculate the collection of expected-by-chance muMCW_BIs.
}
\item It calculates the \emph{P~upper~} and \emph{P~lower~} values as the fraction of expected-by-chance muMCW_BIs that are higher or equal to and lower or equal to the observed muMCW_BIs, respectively.
}
}
}
\examples{
test_temp <- tempdir()
extdata_v <- system.file("extdata", "example_vertical_muMCWtest_data.csv", package = "MCWtests")
file.copy(extdata_v, test_temp)
extdata_h <- system.file("extdata", "example_horizontal_muMCWtest_data.csv", package = "MCWtests")
file.copy(extdata_h, test_temp)
# running muMCWtest with an ideal vertical entry dataset
path_v <- file.path(test_temp, "example_vertical_muMCWtest_data.csv")
muMCWtest_vertical_results <- muMCWtest(path_v, 10)
print(muMCWtest_vertical_results)
# running muMCWtest with an ideal horizontal entry dataset
path_h <- file.path(test_temp, "example_horizontal_muMCWtest_data.csv")
muMCWtest_horizontal_results <- muMCWtest(path_h, 10)
print(muMCWtest_horizontal_results)

rm(test_temp)

}
