% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uMCW_approximated_test.R
\name{uMCW_approximated_test}
\alias{uMCW_approximated_test}
\title{Unmatched-Pairs Monte Carlo-Wilcoxon (uMCW) Approximated Test}
\usage{
uMCW_approximated_test(x)
}
\arguments{
\item{x}{Data for uMCW approximated tests provided by function \emph{uMCWtest}.}
}
\description{
Unmatched-Pairs Monte Carlo-Wilcoxon (uMCW) Approximated Test
}
\keyword{internal}
