% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uMCW_approximated_test_simulated_BIs.R
\name{uMCW_approximated_test_simulated_BIs}
\alias{uMCW_approximated_test_simulated_BIs}
\title{Unmatched-Pairs Monte Carlo-Wilcoxon (uMCW) Approximated Test: Simulated Bias Indexes (BIs)}
\usage{
uMCW_approximated_test_simulated_BIs(x)
}
\arguments{
\item{x}{Data for uMCW approximated tests provided by function \emph{uMCW_approximated_test}.}
}
\description{
Unmatched-Pairs Monte Carlo-Wilcoxon (uMCW) Approximated Test: Simulated Bias Indexes (BIs)
}
\keyword{internal}
