% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uMCW_exact_test.R
\name{uMCW_exact_test}
\alias{uMCW_exact_test}
\title{Unmatched-Pairs Monte Carlo-Wilcoxon (uMCW) Exact Test}
\usage{
uMCW_exact_test(x)
}
\arguments{
\item{x}{Data for uMCW exact tests provided by function \emph{uMCWtest}.}
}
\description{
Unmatched-Pairs Monte Carlo-Wilcoxon (uMCW) Exact Test
}
\keyword{internal}
