% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uMCWtest.R
\name{uMCWtest}
\alias{uMCWtest}
\title{Unmatched-Measures Monte Carlo-Wilcoxon (uMCW) Test}
\format{
When executing the \emph{uMCWtest} function, users must provide the path to a local CSV file named \emph{X_uMCWtest_data.csv}, where \emph{X} serves as a user-defined identifier. \emph{X_uMCWtest_data.csv} can be structured in two distinct formats:
\itemize{
\item \strong{Vertical layout}: This format allows appending datasets with varying structures, such as different numbers of measures per set or between each appended test. Vertical entry datasets should include the following columns:
\itemize{
\item The \emph{condition} column uniquely identifies each of the two measure sets under analysis.
\item The \emph{value} column contains the actual measures under analysis.
\item As many informative columns as needed by users to contextualize the results of each test. The names of these columns should not include the terms \emph{condition} or \emph{value}. While these columns are optional when running a single test, at least one column is required when running multiple tests simultaneously. All rows for each individual test must contain the same information in these columns.
}
\item \strong{Horizontal layout}: This format allows appending datasets with similar structures, such as the same number of measures collected for each of the two conditions. Horizontal entry datasets should include the following columns:
\itemize{
\item Columns \emph{condition_a} and \emph{condition_b} uniquely identify the two measure sets under analysis.
\item Columns \emph{a.i} and \emph{b.j}, where \emph{i} and \emph{j} represent integers to differentiate specific measures within each set, contain the actual measures under analysis.
\item As many informative columns as needed by users to contextualize the results of each test. The names of these columns should not contain the term \emph{condition} or have the same structure as the \emph{a.i} and \emph{b.j} columns. While these columns are optional when running a single test, at least one column is required when running multiple tests simultaneously.
}
}
}
\usage{
uMCWtest(path, max_rearrangements)
}
\arguments{
\item{path}{Path for the local CSV file containing the entry dataset formatted for uMCW tests.}

\item{max_rearrangements}{User-defined maximum number of rearrangements of the dataset used by the function \emph{uMCWtest} to generate a collection of expected-by-chance uMCW_BIs and uMCW_HBIs and estimate the statistical significance of observed uMCW_BIs and uMCW_HBIs. If the number of distinct dataset rearrangements is less than \emph{max_rearrangements}, \emph{uMCWtest} calculates uMCW_BIs and uMCW_HBIs for all possible data rearrangements. If the number of distinct dataset rearrangements is greater than \emph{max_rearrangements}, \emph{uMCWtest} will perform N = \emph{max_rearrangements} random measure rearrangements to calculate the collection of expected-by-chance uMCW_BIs and uMCW_HBIs.}
}
\value{
The \emph{uMCWtest} function reports to the console the total number of tests it will execute, and their exact and approximated counts. It also creates a CSV file named \emph{X_uMCWtest_results.csv}, where \emph{X} is a user-defined identifier for the entry dataset CSV file. The \emph{X_uMCWtest_results.csv} file contains four rows for each uMCWtest, two for uMCW_BIs calculated for each condition contrast (\emph{e.g.}, \emph{a-b} and \emph{b-a}), and two for uMCW_HBIs calculated for each condition contrast. The \emph{X_uMCWtest_results.csv} file includes the following columns:
\itemize{
\item User-provided informative columns to contextualize the results of each test.
\item Columns \emph{condition_a} and \emph{condition_b} indicate the two measure sets under analysis.
\item Columns \emph{N}, \emph{n_a} and \emph{n_b} indicate the total number of measures and the number of measures belonging to each set after removing missing values (NAs).
\item Column \emph{test_type} distinguishes between exact and approximated tests.
\item Column \emph{BI_type} indicates the bias index type (uMCW_BI and uMCW_HBI) for each row of results.
\item Column \emph{condition_contrast} indicates the set contrast (\emph{e.g.}, \emph{a-b} or \emph{b-a}) for each row of results.
\item Column \emph{observed_BI} contains the values of uMCW_BIs and uMCW_HBIs obtained from analyzing the user-provided dataset.
\item Column \emph{expected_by_chance_BI_N} indicates the number of data rearrangements used to calculate the expected-by-chance uMCW_BIs and uMCW_HBIs. This value corresponds to the lowest number between all possible measure rearrangements and the parameter \emph{max_rearrangements}.
\item Columns \emph{pupper} and \emph{plower} represent the \emph{P~upper~} and \emph{P~lower~} values, respectively. They denote the fraction of expected-by-chance uMCW_BIs or uMCW_HBIs with values higher or equal to and lower or equal to the observed uMCW_BIs or uMCW_HBIs, respectively.
}
}
\description{
The uMCW test is a statistical tool to assess whether two sets of unmatched measures and their heterogeneity are significantly biased in the same direction. Significantly different data heterogeneities between two conditions could indicate that the measure under analysis is more constrained or more relaxed in one of the conditions, potentially providing insights into the mechanisms underlying the variation of such measure. For instance, uMCW tests can be used to analyze bodyweights or transcript abundances determined for two sets of mice that have been maintained in different conditions.
}
\details{
The function \emph{uMCWtest} eliminate missing values (NAs) from the dataset before proceeding these steps.
\itemize{
\item To estimate the bias between the two sets of measures (\emph{e.g.}, \emph{a} and \emph{b}), the function \emph{uMCWtest} performs these tasks:
\itemize{
\item It generates all possible disjoint data pairs using measures from both sets.
\item For each measure pair, it subtracts the second measure in the pair from the first measure in the pair.
\item It ranks the absolute values of all non-zero measure pair differences from lowest to highest. Measure pair differences with a value of 0 are assigned a 0 rank. If multiple measure pair differences have the same absolute value, all tied measure pair differences are assigned the lowest rank possible.
\item It assigns each measure pair rank a sign based on the sign of its corresponding measure pair difference.
\item It sums the signed ranks for measure pairs formed with measures from the two different sets (\emph{e.g.}, \emph{a-b} and \emph{b-a}).
\item For each type of disjoint set measure pairs (\emph{e.g.}, \emph{a-b} and \emph{b-a}), it calculates uMCW_BI by dividing the sum of signed ranks by the maximum number this sum could have if the corresponding measure pairs had the highest possible positive ranks. Consequently, uMCW_BI ranges between 1 when all the values for measures in the first set are higher than all the values from measures in the second set, and -1 when all the values for measures in the first set are lower than all the values from measures in the second set.
}
\item To estimate the bias between the heterogeneity of two sets of measures, the function \emph{uMCWtest} performs these tasks:
\itemize{
\item It generates all possible disjoint data pairs within each set, disregarding the order of the paired measures. For instance, the measure pair \emph{a.1-a.2} is considered equivalent to the measure pair \emph{a.2-a.1}, and only the former is retained for the subsequent calculations.
\item For each measure pair, it subtracts the second measure from the first measure.
\item It ranks all measure pair differences with non-zero values from lowest to highest. Measure pair differences with a value of 0 are assigned a 0 rank. If multiple measure pair differences have the same absolute value, \emph{uMCWtest} assigns all tied measure pair differences the lowest rank possible.
\item It sums ranks for measure pairs formed with measures from the same set (\emph{e.g.}, \emph{a-a} and \emph{b-b}).
\item For each type of same-set measure pairs (\emph{e.g.}, \emph{a-a} and \emph{b-b}), it divides each sum of signed ranks by the maximum number this sum could have if the corresponding measure pairs had the highest possible ranks.
\item It calculates two heterogeneity bias indexes (uMCW_HBIs) by subtracting the normalized sum of signed ranks from the previous step in two possible directions (\emph{e.g.}, \emph{a-b} and \emph{b-a}). Consequently, uMCW_HBI ranges between 1 when at least two measures in the first set have distinct values and all measures in the second set have the same value, and -1 when all measures in the first set have the same value and at least two measures in the second set have distinct values.
}
\item To assess the significance of the uMCW_BIs and uMCW_HBIs obtained with the user-provided data (observed uMCW_BIs and uMCW_HBIs), the function \emph{uMCWtest} performs these tasks:
\itemize{
\item It generates a collection of expected-by-chance uMCW_BIs and uMCW_HBIs. These expected values are obtained by rearranging the measures between the two sets multiple times. The user-provided parameter \emph{max_rearrangements} determines the two paths that the function \emph{uMCWtest} can follow to generate the collection of expected-by-chance uMCW_BIs and uMCW_HBIs:
\itemize{
\item \emph{uMCW exact testing}: If the number of distinct measure rearrangements that can alter their initial set distribution is less than \emph{max_rearrangements}, the function \emph{uMCWtest} calculates uMCW_BIs and uMCW_HBIs for all possible data rearrangements.
\item \emph{uMCW approximated testing}: If the number of distinct measure rearrangements that can alter their initial set distribution is greater than \emph{max_rearrangements}, the function \emph{uMCWtest} will perform N = \emph{max_rearrangements} random measure rearrangements to calculate the collection of expected-by-chance uMCW_BIs and uMCW_HBIs.
}
\item It calculates \emph{P~upper~} and \emph{P~lower~} values as the fraction of expected-by-chance uMCW_BIs and uMCW_HBIs that are higher or equal to and lower or equal to the observed uMCW_BIs and uMCW_HBIs, respectively.
}
}
}
\examples{
# Executing uMCWtest with an ideal vertical entry dataset
path_v <- system.file("extdata", "example_vertical_uMCWtest_data.csv", package = "MCWtests")
uMCWtest_v_results <- uMCWtest(path_v, 200)

# Executing uMCWtest with an ideal horizontal entry dataset
path_h <- system.file("extdata", "example_horizontal_uMCWtest_data.csv", package = "MCWtests")
uMCWtest_h_results <- uMCWtest(path_h, 200)

}
